pyproject.toml
[tool.poetry]
name = "expl-test"
version = "0.1.0"
description = "тестТестовое задание"
authors = ["AlexeyRybicky <Ranangariel@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydentic = {extras = ["email"], version = "^0.0.1.dev3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.21"}
aiosqlite = "^0.19.0"
pydantic-settings = "^2.0.3"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pytest = "^7.4.2"
logging = "^0.4.9.6"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    '.', 'fastapi_app'
]